Implement s2c12
Bit of a hack - when we hit pkcs7 padding at the end of the phrase,
we'll start getting no matches on target block (because the padding
was \001, then would be \002\002 on the next byte etc). Just stop in
that case. Could have done something more clever by checking whether
the last bytes look like padding and chop them all off but meh.